Remove coloured diff output
The point of outputing difs to stderr was to give the user the flexibility to
redirect stderr to their output file or something else similar. As we are
using coloring codes for BASH the redirected output would contain these bytes
as in and text editors like gedit/vim are not smart enough to parse these
colors.